moved functions into a components-folder
