+++ src/js/templates/components.mjs
@@ -1,0 +1,21 @@
+export function getQueryStringParam(param) {
+  const urlParams = new URLSearchParams(window.location.search);
+  return urlParams.get(param);
+}
+
+export function formatDateString(dateString) {
+  const date = new Date(dateString);
+
+  const options = {
+    year: "numeric",
+    month: "long",
+    day: "numeric",
+    hour: "numeric",
+    minute: "numeric",
+    hour12: true,
+  };
+
+  const formattedDate = date.toLocaleString("en-US", options);
+
+  return formattedDate;
+}
